Version 1.4
Made to work with split window

Important Bug fix.(Fixed the disappearing of autocommands the function list window in another tab is closed)

Made recent window to update on navigating to a function using it.
Fixed a bug where the function list window is still in insert mode, so double click wont take you to the function
Added auto update of recent function list while entering or leaving insert mode.